managing users

+++ app/Http/Controllers/CategoriesController.php
@@ -101,6 +101,12 @@ class CategoriesController extends Controller
      */
     public function destroy(Category $category)
     {
+        if ($category->posts->count() > 0){
+
+            session()->flash('error', 'Category cannot be deleted because it has some post.');
+            return redirect()->back();
+        }
+
         $category->delete();
 
         session()->flash('success', 'Category deleted successfully');
